Fixed some bugs in pebble.py
Lines 16 and 17 gave the values [False] and [-1] to pebble[str(i)] and
pebble_value[str(i)] instead of False and -1, which broke the
add_pebble() function.  It is now working again.  This bug took me a
while to find because I kept looking for problems in the add_pebble()
function but it ended up being in the constructor.  Oops.

Will try to debug pebbling_algos.py next.  John please see my comment on
line 31 of pebble.py in the remove_pebble() function.