Hmm ok so apparently pebble.py is still not entirely fixed
Hopefully we can talk about this at the meeting -- why do pebble and
pebble_value need to be dictionaries and not arrays?  Is that just how
shelf works or something?  In that case we are going to have to
implement some more workarounds.

Also why is pebble.txt a 10MB file for the r=2 PTC graph (24 vtxs)
:astonished: